refactor: move things from common that are only used in ecr document
